Ya puedes buscar a los usuarios y mandar solicitud de amistad, solo se puede entrar en los perfiles de tus amigos, si no es amigo tuyo no te dejará entrar. Hemos empezado a implementar el chat.

+++ src/components/FriendInfoComponent.vue
@@ -5,37 +5,61 @@ import { RouterLink } from "vue-router";
 export default {
   name: "FriendInfo",
   props: {
-    friend: {
+    friendId: {
       type: Object,
       required: true
     }
   },
   data() {
     return {
-      followText: "Unfollow"
+      friend: {},
     }
   },
-
+  created() {
+    if (localStorage.getItem("token")) {
+      this.getFriendInfo();
+    } else {
+      router.push({ name: "Login" });
+    }
+  },
   methods: {
-    unfollowFriend() {
-      fetch('https://balandrau.salle.url.edu/i3/socialgift/api/v1/friends/'+ this.friend.id, {
-        method: 'DELETE',
+    getFriendInfo() {
+      fetch('https://balandrau.salle.url.edu/i3/socialgift/api/v1/users/' + this.friendId, {
         headers: {
           "accept": "application/json",
           "Authorization": 'Bearer ' + localStorage.getItem("token"),
           "Content-Type": 'application/json'
         }
       })
-        .then(response => {
-          if (response.ok) {
-            return response.json()
-          } else {
-            throw new Error(response.statusText)
+        .then((data) => data.json())
+        .then((json) => {
+          this.friend = json;
+        })
+        .catch((error) => {
+          console.log("error: " + error);
+        });
+    },
+    unfollowFriend() {
+      if (confirm("Are you sure you want to unfollow "+this.friend.name +" "+this.friend.last_name+"?")) {
+        fetch('https://balandrau.salle.url.edu/i3/socialgift/api/v1/friends/' + this.friend.id, {
+          method: 'DELETE',
+          headers: {
+            "accept": "application/json",
+            "Authorization": 'Bearer ' + localStorage.getItem("token"),
+            "Content-Type": 'application/json'
           }
         })
-        .catch(error => {
-          console.log("error: " + error)
-        })
+          .then(response => {
+            if (response.ok) {
+              router.push({ name: "MyUser" })
+            } else {
+              throw new Error(response.statusText)
+            }
+          })
+          .catch(error => {
+            console.log("error: " + error)
+          })
+      }
     }
   }
 }
@@ -58,7 +82,7 @@ export default {
     <!--Componente botones del amigo-->
     <div class="friend-buttons-div">
       <router-link class="message-button" to="/Messages">Message</router-link>
-      <button class="unfollow-button" @click="unfollowFriend()">{{followText}}</button>
+      <button class="unfollow-button" @click="unfollowFriend()">Unfollow</button>
     </div>
   </div>
 </template>
